+++ src/GameManager.cpp
@@ -1,10 +1,7 @@
-#pragma once
-
 #include "EditorUI.hpp"
 #include <Geode/modify/GameManager.hpp>
 
 class $modify(GameManager) {
-    
     gd::string stringForCustomObject(int customObjectID) {
         if (customObjectID <= CUSTOM_OBJECT_ID_OFFSET) {
             if (auto editor = Global::editor()) {