+++ changelog.md
@@ -1,3 +1,9 @@
+# 2.1.2
+- Dragging groups no longer breaks with mods that use ImGui (e.g DevTools)
+- Opened groups no longer closed when switching to 'Edit Groups' mode
+- '<cj>Object Search</c>' feature - open the tab and group that contain selected object
+
+
 # 2.1.0
 - Group pinning:
   - <cp>Pinning</c> groups so that they stay on the screen